chore: refactored api/routers/applications
separated concerns in different files extracting the logic of fetching applications and approvals from attestations to be reused by other routers in the server side
NOTE: doesn't affect the codebase as everything exported previously by applications.ts is exported now by applications/index.ts
