+++ src/server/api/routers/applications/index.ts
@@ -1,0 +1,1 @@
+export { applicationsRouter, FilterSchema } from "./applicationsRouter";
